+++ src/main/webapp/WEB-INF/views/movimentacao/movimentacaopedidovenda.jsp
@@ -23,7 +23,7 @@
           
           
           <div class="widget-content nopadding">
-            <table class="table table-bordered data-table">
+            <table class="table table-hover table-bordered">
               <thead>
                <tr>
                                         <th>Id</th>
@@ -84,15 +84,25 @@
 									title="Informações" ><i class="icon-info-sign"></i></a>
 									
 									
+									<a href="#myAlert${pedidovenda.id}" data-toggle="modal" class="fa fa-remove"><i class="icon-remove-sign"></i></a>
 									
+			<div id="myAlert${pedidovenda.id}" class="modal hide">
+              <div class="modal-header">
+                <button data-dismiss="modal" class="close" type="button">×</button>
+                <h3>Alerta de Exclusão</h3>
+              </div>
+              <div class="modal-body">
+                <p>Deseja Realmente Excluir esse Registro</p>
+              </div>
+              <div class="modal-footer"> <a data-dismiss="" class="btn btn-danger" href="${pageContext.request.contextPath}/pedidovenda/delete?id=${pedidovenda.id}">Confirma</a> <a data-dismiss="modal" class="btn" href="#">Cancela</a> </div>
+           
+            </div>
+
 									
-									 <a
-									href="${pageContext.request.contextPath}/pedidovenda/delete?id=${pedidovenda.id}"
-									title="deletar" class="fa fa-remove fa-2x" ><i class="icon-minus-sign"></i> </a>
-									
-									
-									
-									
+<!-- 									 <a -->
+<%-- 									href="${pageContext.request.contextPath}/pedidovenda/delete?id=${pedidovenda.id}" --%>
+<!-- 									title="deletar" class="fa fa-remove fa-2x" ><i class="icon-minus-sign"></i> </a> -->
+											
 									
 										
 					</td>
@@ -104,6 +114,9 @@
                 </c:forEach>
 
               </tbody>
+              
+
+              
             </table>
           </div>
         </div>
@@ -112,9 +125,12 @@
         </div>
         
 <!--         <div class="pagination alternate"> </div> -->
-        
+
         
         </div>
+
+
+          
         
         <div class="row-fluid">
   <div id="footer" class="span12"> 2017 &copy; Tecsoluction LTDA <a href="http://themedesigner.in">Soluções em Tecnologia</a> </div>
