+++ src/main/webapp/WEB-INF/views/movimentacao/movimentacaopedidovendaentregas.jsp
@@ -23,7 +23,7 @@
           
           
           <div class="widget-content nopadding">
-            <table class="table table-bordered data-table">
+            <table class="table table-hover table-bordered">
               <thead>
                <tr>
                                         <th>Id</th>
@@ -81,11 +81,28 @@
 										<a
 									href="${pageContext.request.contextPath}/pedidovenda/editar?id=${pedidovenda.id}"
 									title="editar" class="fa fa-pencil fa-2x"></a>
+	
+	
+	
+	
+	
+		<a href="#myAlert${pedidovenda.id}" data-toggle="modal" class="fa fa-remove"><i class="icon-remove-sign"></i></a>
 									
+			<div id="myAlert${pedidovenda.id}" class="modal hide">
+              <div class="modal-header">
+                <button data-dismiss="modal" class="close" type="button">×</button>
+                <h3>Alerta de Exclusão</h3>
+              </div>
+              <div class="modal-body">
+                <p>Deseja Realmente Excluir esse Registro</p>
+              </div>
+              <div class="modal-footer"> <a data-dismiss="" class="btn btn-danger" href="${pageContext.request.contextPath}/pedidovenda/delete?id=${pedidovenda.id}">Confirma</a> <a data-dismiss="modal" class="btn" href="#">Cancela</a> </div>
+           
+            </div>								
 									
-									 <a
-									href="${pageContext.request.contextPath}/pedidovenda/delete?id=${pedidovenda.id}"
-									title="deletar" class="fa fa-remove fa-2x" ></a>
+<!-- 									 <a -->
+<%-- 									href="${pageContext.request.contextPath}/pedidovenda/delete?id=${pedidovenda.id}" --%>
+<!-- 									title="deletar" class="fa fa-remove fa-2x" ></a> -->
